Move texture loading to it's own module
